requirements.txt
Django==2.2.6
django-cors-headers==3.3.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
importlib-metadata==1.6.1
Markdown==3.2.2
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
zipp==3.1.0
